Remove k256 dependency and unused code
