+++ raiko-guest/src/signature.rs
@@ -18,6 +18,7 @@ pub fn generate_key() -> KeyPair {
 ///
 /// This does not ensure that the `s` value in the signature is low, and _just_ wraps the
 /// underlying secp256k1 library.
+#[allow(dead_code)]
 pub fn recover_signer_unchecked(sig: &[u8; 65], msg: &[u8; 32]) -> Result<Address, Error> {
     let sig =
         RecoverableSignature::from_compact(&sig[0..64], RecoveryId::from_i32(sig[64] as i32)?)?;
@@ -58,4 +59,30 @@ pub fn load_private_key<T: AsRef<Path>>(path: T) -> Result<SecretKey, Error> {
 
 pub fn public_key(secret: &SecretKey) -> PublicKey {
     PublicKey::from_secret_key(&Secp256k1::new(), secret)
+}
+
+#[cfg(test)]
+mod tests {
+    use std::str::FromStr;
+
+    use super::*;
+    #[test]
+    fn recover() {
+        let proof = "01000000c13bd882edb37ffbabc9f9e34a0d9789633b850fe55e625b768cc8e5feed7d9f7ab536cbc210c2fcc1385aaf88d8a91d8adc2740245f9deee5fd3d61dd2a71662fb6639515f1e2f3354361a82d86c1952352c1a800";
+        let proof_bytes = hex::decode(proof).unwrap();
+        let msg = "216ac5cd5a5e13b0c9a81efb1ad04526b9f4ddd2fe6ebc02819c5097dfb0958c";
+        let msg_bytes = hex::decode(msg).unwrap();
+        let proof_addr = recover_signer_unchecked(
+            &proof_bytes[24..].try_into().unwrap(),
+            &msg_bytes.try_into().unwrap(),
+        )
+        .unwrap();
+        let priv_key = "324b5d1744ec27d6ac458350ce6a6248680bb0209521b2c730c1fe82a433eb54";
+        let priv_key = SecretKey::from_str(priv_key).unwrap();
+        let pubkey = public_key(&priv_key);
+        let pub_addr = public_key_to_address(pubkey);
+        assert_eq!(pub_addr, proof_addr);
+        println!("Public address: {}", pub_addr);
+        println!("Proof public address: {}", proof_addr);
+    }
 }
